tutorials/predict-and-approve/requirements.txt
lucidtech-las >= 3.0.0
